fix: 렌더링마다 유저 정보가져와서 권한 확인하도록 수정
